requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
auto_setup==0.1.1
bcrypt==4.2.1
certifi==2024.12.14
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.6
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.2
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.5
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
